Moved LoginSignupComponents into pages folder also updating the import statements
